image-bridge/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
mflux>=0.9.0
Pillow>=10.0.0
httpx>=0.25.0
